Add files via upload
Chinese Zodiac Year Calculator
This Python project calculates the Chinese zodiac year for a given year. The Chinese zodiac is a repeating cycle of 12 years, with each year being represented by an animal and its reputed attributes.
How to Use
Run the Program: Execute the Python script to start the calculator.
Input the Year: Enter the year for which you want to find the zodiac sign.
Get the Result: The program will display the corresponding Chinese zodiac animal for that year.