Add files via upload
In the provided Python code snippets, the program retrieves and prints the current local time using a time module. With the time module, it imports time, then uses time.time() to get the current time as a structured time object. It formats this time using str() for covertion for the time to be  return to a string in the HH:MM:SS format and prints the result.
To get the total number of seconds , the current time is divide by total number of seconds, and the current seconds at that moment is gotten by the modulo total number of seconds. same method is applied for the rest of minutes and hours.